+++ demo/src/components/controls/calendar.tsx
@@ -36,7 +36,7 @@ export default function CalendarControlView() {
         ref={(el) => {
           if (el) {
             el.$initModel = { readonly: new Date("1990-02-10") };
-            el.$onSubmit = (e) => console.warn("submitted model", e.$model);
+            el.$onSubmit = (e) => console.warn("submitted model", e.detail.model);
           }
         }}
         w-autoFocus