+++ demo/src/components/controls/text.tsx
@@ -34,7 +34,7 @@ export default function TextControlView() {
         ref={(el) => {
           if (el) {
             el.$initModel = { email: "test@google.com" };
-            el.$onSubmit = (e) => console.warn("submitted model", e.$model);
+            el.$onSubmit = (e) => console.warn("submitted model", e.detail.model);
           }
         }}
         w-autoFocus