+++ src/formElement.ts
@@ -24,24 +24,26 @@ export const enum SubmitActions {
 const tagName = "wup-form";
 declare global {
   namespace WUP.Form {
-    interface SubmitEvent<T extends Record<string, any>> extends Event {
+    interface SubmitDetails {
       /** Model collected from controls */
-      $model: Partial<T>;
-      /** Form related to submit event */
-      $relatedForm: WUPFormElement<T>;
+      model: Record<string | number, any>;
+      /** Form related to submit event; equal to event.target */
+      relatedForm: WUPFormElement;
       /** Event that produced submit event */
-      $relatedEvent: MouseEvent | KeyboardEvent;
+      relatedEvent: MouseEvent | KeyboardEvent;
       /** Element that that produced submit event */
-      $submitter: HTMLElement | null;
+      submitter: HTMLElement | null;
       /** Point a promise as callback to allow form show pending state during the promise */
-      $waitFor?: Promise<unknown>;
+      waitFor?: Promise<unknown>;
     }
 
     interface EventMap extends WUP.Base.EventMap {
-      /** Fires before $submit is happened; can be prevented via e.preventDefault() */
-      $willSubmit: Omit<SubmitEvent<any>, "$model">;
+      /** Fires before $submit is happened; can be prevented via `e.preventDefault()` */
+      $willSubmit: CustomEvent<Pick<SubmitDetails, "relatedEvent" | "relatedForm" | "submitter">>;
       /** Fires by user-submit when validation succesfull and model is collected */
-      $submit: SubmitEvent<any>;
+      $submit: CustomEvent<SubmitDetails>;
+      /** Fires when submit is end (after http-response) */
+      $submitEnd: CustomEvent<{ success: boolean }>;
     }
 
     interface Options {
@@ -113,8 +115,8 @@ const formStore: WUPFormElement[] = [];
  *  const form = document.createElement("wup-form");
  *  form.$options.autoComplete = false;
  *  form.$initModel = { email: "test-me@google.com" };
- *  form.addEventListener("$submit", (e) => console.warn(e.$model) );
- *  form.$onSubmit = async (e)=>{ await postHere(e.$model); } // equal to form.addEventListener
+ *  form.addEventListener("$submit", (e) => console.warn(e.detail.model) );
+ *  form.$onSubmit = async (e)=>{ await postHere(e.detail.model); } // equal to form.addEventListener
  *  // init control
  *  const el = document.createElement("wup-text");
  *  el.$options.name = "email";
@@ -259,8 +261,12 @@ export default class WUPFormElement<
     readOnly: false,
   };
 
-  /** Dispatched on submit. Return promise to lock form and show spinner */
-  $onSubmit?: (ev: WUP.Form.SubmitEvent<Model>) => void | Promise<unknown>;
+  /** Fires before $submit is happened; can be prevented via `e.preventDefault()` */
+  $onWillSubmit?: (ev: WUP.Form.EventMap["$willSubmit"]) => void;
+  /** Dispatched on submit. Return promise to lock form and show spinner on http-request */
+  $onSubmit?: (ev: WUP.Form.EventMap["$submit"]) => void | Promise<unknown>;
+  /** Fires when submit is end (after http-response) */
+  $onSubmitEnd?: (ev: WUP.Form.EventMap["$submitEnd"]) => void;
   /** Dispatched on submit */
   // It's not required but called: $onsubmit?: (ev: WUP.Form.SubmitEvent<Model>) => void;
 
@@ -366,15 +372,17 @@ export default class WUPFormElement<
 
   /** Called on submit before validation */
   protected gotSubmit(e: KeyboardEvent | MouseEvent, submitter: HTMLElement): void {
-    e.preventDefault();
+    e.preventDefault(); // prevent default keyboard or mouse event because it's handled in custom event
 
-    const willEv = new Event("$willSubmit", { bubbles: true, cancelable: true }) as Events["$willSubmit"];
-    // willEv.$model = null;
-    willEv.$relatedEvent = e;
-    willEv.$relatedForm = this as WUPFormElement<any>;
-    willEv.$submitter = submitter;
-
-    this.dispatchEvent(willEv);
+    const willEv = this.fireEvent("$willSubmit", {
+      bubbles: true,
+      cancelable: true,
+      detail: {
+        relatedEvent: e,
+        relatedForm: this,
+        submitter,
+      },
+    });
     if (willEv.defaultPrevented) {
       return;
     }
@@ -409,16 +417,21 @@ export default class WUPFormElement<
     const onlyChanged = this._opts.submitActions & SubmitActions.collectChanged;
     const m = this.#ctr.$modelFromControls({}, arrCtrl, "$value", !!onlyChanged);
     // fire events
-    const ev = new Event("$submit", { cancelable: false, bubbles: true }) as WUP.Form.SubmitEvent<Model>;
-    ev.$model = m;
-    ev.$relatedForm = this;
-    ev.$relatedEvent = e;
-    ev.$submitter = submitter;
+    const ev: WUP.Form.EventMap["$submit"] = new CustomEvent("$submit", {
+      cancelable: false,
+      bubbles: true,
+      detail: {
+        model: m,
+        relatedEvent: e,
+        relatedForm: this,
+        submitter,
+      },
+    });
 
     const needReset = this._opts.submitActions & SubmitActions.reset;
     setTimeout(() => {
-      this.dispatchEvent(ev);
       const p1 = this.$onSubmit?.call(this, ev);
+      this.dispatchEvent(ev);
       // SubmitEvent constructor doesn't exist on some browsers: https://developer.mozilla.org/en-US/docs/Web/API/SubmitEvent/SubmitEvent
       const ev2 = new (window.SubmitEvent || Event)("submit", { submitter, cancelable: false, bubbles: true });
       /* istanbul ignore else */
@@ -427,13 +440,19 @@ export default class WUPFormElement<
       }
       this.dispatchEvent(ev2);
 
-      promiseWait(Promise.all([p1, ev.$waitFor]), 300, (v: boolean) => this.changePending(v)).then(() => {
-        this._opts.autoStore && this.storageSave(null); // clear storage after submit
-        if (needReset) {
-          arrCtrl.forEach((v) => (v.$isDirty = false));
-          this.$initModel = this.$model;
-        }
-      });
+      let success = false;
+      promiseWait(Promise.all([p1, ev.detail.waitFor]), 300, (v: boolean) => this.changePending(v))
+        .then(() => {
+          this._opts.autoStore && this.storageSave(null); // clear storage after submit
+          if (needReset) {
+            arrCtrl.forEach((v) => (v.$isDirty = false));
+            this.$initModel = this.$model;
+          }
+          success = true;
+        })
+        .finally(() => {
+          this.fireEvent("$submitEnd", { detail: { success }, cancelable: false, bubbles: true });
+        });
     });
   }